Correctly handle large UIDs/GIDs as float64

+++ providers/os/connection/local/local_unix.go
@@ -13,16 +13,16 @@ import (
 	"go.mondoo.com/cnquery/v11/providers/os/connection/shared"
 )
 
-func (c *LocalConnection) fileowner(stat os.FileInfo) (int64, int64) {
-	uid := int64(-1)
-	gid := int64(-1)
+func (c *LocalConnection) fileowner(stat os.FileInfo) (float64, float64) {
+	uid := float64(-1)
+	gid := float64(-1)
 	if stat, ok := stat.Sys().(*shared.FileInfo); ok {
-		uid = stat.Uid
-		gid = stat.Gid
+		uid = float64(stat.Uid)
+		gid = float64(stat.Gid)
 	}
 	if stat, ok := stat.Sys().(*syscall.Stat_t); ok {
-		uid = int64(stat.Uid)
-		gid = int64(stat.Gid)
+		uid = float64(stat.Uid)
+		gid = float64(stat.Gid)
 	}
 	return uid, gid
 }
