+++ providers/os/resources/groups/etcgroups_test.go
@@ -30,13 +30,13 @@ func TestParseLinuxEtcGroups(t *testing.T) {
 
 	assert.Equal(t, "root", m[0].Name, "detected user name")
 	assert.Equal(t, "0", m[0].ID, "detected id")
-	assert.Equal(t, int64(0), m[0].Gid, "detected gid")
+	assert.Equal(t, float64(0), m[0].Gid, "detected gid")
 	assert.Equal(t, "", m[22].Sid, "detected sid")
 	assert.Equal(t, []string{}, m[0].Members, "user description")
 
 	assert.Equal(t, "vagrant", m[22].Name, "detected user name")
 	assert.Equal(t, "1000", m[22].ID, "detected id")
-	assert.Equal(t, int64(1000), m[22].Gid, "detected gid")
+	assert.Equal(t, float64(1000), m[22].Gid, "detected gid")
 	assert.Equal(t, "", m[22].Sid, "detected sid")
 	assert.Equal(t, []string{"vagrant"}, m[22].Members, "user description")
 }
@@ -59,13 +59,13 @@ func TestParseFreebsd12EtcGroups(t *testing.T) {
 
 	assert.Equal(t, "wheel", m[0].Name, "detected user name")
 	assert.Equal(t, "0", m[0].ID, "detected id")
-	assert.Equal(t, int64(0), m[0].Gid, "detected gid")
+	assert.Equal(t, float64(0), m[0].Gid, "detected gid")
 	assert.Equal(t, "", m[0].Sid, "detected sid")
 	assert.Equal(t, []string{"root", "vagrant"}, m[0].Members, "user description")
 
 	assert.Equal(t, "vagrant", m[35].Name, "detected user name")
 	assert.Equal(t, "1001", m[35].ID, "detected id")
-	assert.Equal(t, int64(1001), m[35].Gid, "detected gid")
+	assert.Equal(t, float64(1001), m[35].Gid, "detected gid")
 	assert.Equal(t, "", m[35].Sid, "detected sid")
 	assert.Equal(t, []string{}, m[35].Members, "user description")
 }
@@ -94,7 +94,7 @@ func TestUnixGroupManager(t *testing.T) {
 
 	require.NotNil(t, vagrantGroup)
 	assert.Equal(t, "vagrant", vagrantGroup.Name)
-	assert.Equal(t, int64(1000), vagrantGroup.Gid)
+	assert.Equal(t, float64(1000), vagrantGroup.Gid)
 	assert.Equal(t, "1000", vagrantGroup.ID)
 	assert.Equal(t, []string{"vagrant"}, vagrantGroup.Members)
 }
